Fixed visibility of view menu when starting with a selected data source Changed main menu items to outline variant, where possible Search and find all text box are hidden when no data source is selected
