+++ src/Installer/packages.config
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="log4net" version="2.0.12" targetFramework="net471" />
-  <package id="Metrolib" version="0.3.0.159" targetFramework="net471" />
+  <package id="Metrolib" version="0.3.0.160" targetFramework="net471" />
 </packages>
