+++ src/Tailviewer.AcceptanceTests/packages.config
@@ -4,7 +4,7 @@
   <package id="FluentAssertions" version="5.6.0" targetFramework="net471" />
   <package id="FluentAssertions.Extensions" version="1.0.18" targetFramework="net471" />
   <package id="log4net" version="2.0.12" targetFramework="net471" />
-  <package id="Metrolib" version="0.3.0.159" targetFramework="net471" />
+  <package id="Metrolib" version="0.3.0.160" targetFramework="net471" />
   <package id="Moq" version="4.7.127" targetFramework="net471" />
   <package id="NUnit" version="3.12.0" targetFramework="net471" />
   <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.2" targetFramework="net471" />
