+++ src/Tailviewer.Archiver/packages.config
@@ -2,7 +2,7 @@
 <packages>
   <package id="CommandLineParser" version="2.5.0" targetFramework="net452" />
   <package id="log4net" version="2.0.12" targetFramework="net452" />
-  <package id="Metrolib" version="0.3.0.159" targetFramework="net452" />
+  <package id="Metrolib" version="0.3.0.160" targetFramework="net452" />
   <package id="PE" version="0.1.13" targetFramework="net452" />
   <package id="SharpRemote" version="0.6.184" targetFramework="net452" />
   <package id="System.Threading.Extensions" version="2.0.59" targetFramework="net452" />
