+++ src/Tailviewer.Core/packages.config
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="log4net" version="2.0.12" targetFramework="net45" />
-  <package id="Metrolib" version="0.3.0.159" targetFramework="net45" />
+  <package id="Metrolib" version="0.3.0.160" targetFramework="net45" />
   <package id="System.Threading.Extensions" version="2.0.59" targetFramework="net45" />
 </packages>
