+++ src/Tailviewer/App.config
@@ -11,7 +11,7 @@
       </dependentAssembly>
       <dependentAssembly>
         <assemblyIdentity name="metrolib" publicKeyToken="951c63ca832f7094" culture="neutral" />
-        <bindingRedirect oldVersion="0.3.0.0-9999.9999.9999.9999" newVersion="0.3.0.159" />
+        <bindingRedirect oldVersion="0.3.0.0-9999.9999.9999.9999" newVersion="0.3.0.160" />
       </dependentAssembly>
       <dependentAssembly>
         <assemblyIdentity name="System.Extensions" publicKeyToken="951c63ca832f7094" culture="neutral" />
