+++ src/Tailviewer/Ui/DataSourceTree/FolderDataSourceViewModel.cs
@@ -11,8 +11,8 @@ using Tailviewer.BusinessLogic.ActionCenter;
 using Tailviewer.BusinessLogic.DataSources;
 using Tailviewer.BusinessLogic.FileExplorer;
 using Tailviewer.Settings;
-using Tailviewer.Ui.ContextMenu;
 using Tailviewer.Ui.LogView;
+using Tailviewer.Ui.Menu;
 
 namespace Tailviewer.Ui.DataSourceTree
 {
