+++ src/Tailviewer/Ui/DataSourceTree/IDataSourceViewModel.cs
@@ -5,8 +5,8 @@ using System.Windows.Input;
 using Metrolib;
 using Tailviewer.Archiver.Plugins.Description;
 using Tailviewer.BusinessLogic.DataSources;
-using Tailviewer.Ui.ContextMenu;
 using Tailviewer.Ui.LogView;
+using Tailviewer.Ui.Menu;
 
 namespace Tailviewer.Ui.DataSourceTree
 {
