+++ src/Tailviewer/Ui/DataSourceTree/MergedDataSourceViewModel.cs
@@ -12,8 +12,8 @@ using Tailviewer.BusinessLogic.ActionCenter;
 using Tailviewer.BusinessLogic.DataSources;
 using Tailviewer.BusinessLogic.FileExplorer;
 using Tailviewer.Settings;
-using Tailviewer.Ui.ContextMenu;
 using Tailviewer.Ui.LogView;
+using Tailviewer.Ui.Menu;
 
 namespace Tailviewer.Ui.DataSourceTree
 {
