+++ src/Tailviewer/Ui/Menu/ExcludeAllInGroupViewModel.cs
@@ -4,26 +4,26 @@ using System.Windows.Media;
 using Metrolib;
 using Tailviewer.Ui.DataSourceTree;
 
-namespace Tailviewer.Ui.ContextMenu
+namespace Tailviewer.Ui.Menu
 {
-	public sealed class IncludeAllInGroupViewModel
+	public sealed class ExcludeAllInGroupViewModel
 		: IMenuViewModel
 	{
 		private readonly IMergedDataSourceViewModel _viewModel;
 		private readonly DelegateCommand2 _command;
 
-		public IncludeAllInGroupViewModel(IMergedDataSourceViewModel viewModel)
+		public ExcludeAllInGroupViewModel(IMergedDataSourceViewModel viewModel)
 		{
 			_viewModel = viewModel;
-			_command = new DelegateCommand2(OnIncludeAll);
+			_command = new DelegateCommand2(OnExcludeAll);
 		}
 
-		private void OnIncludeAll()
+		private void OnExcludeAll()
 		{
 			foreach (var dataSource in _viewModel.Observable)
 			{
 				if (dataSource is ISingleDataSourceViewModel singleDataSource)
-					singleDataSource.ExcludeFromParent = false;
+					singleDataSource.ExcludeFromParent = true;
 			}
 		}
 
@@ -31,7 +31,7 @@ namespace Tailviewer.Ui.ContextMenu
 
 		public string Header
 		{
-			get { return "Include all"; }
+			get { return "Exclude all"; }
 		}
 
 		public string ToolTip => null;
