+++ src/Tailviewer/Ui/Menu/IMenuViewModel.cs
@@ -2,7 +2,7 @@
 using System.Windows.Input;
 using System.Windows.Media;
 
-namespace Tailviewer.Ui.ContextMenu
+namespace Tailviewer.Ui.Menu
 {
 	public interface IMenuViewModel
 	{
