+++ src/Tailviewer/packages.config
@@ -2,7 +2,7 @@
 <packages>
   <package id="Extended.Wpf.Toolkit" version="4.0.1" targetFramework="net471" />
   <package id="log4net" version="2.0.12" targetFramework="net471" />
-  <package id="Metrolib" version="0.3.0.159" targetFramework="net471" />
+  <package id="Metrolib" version="0.3.0.160" targetFramework="net471" />
   <package id="MMQ" version="1.0.15" targetFramework="net471" />
   <package id="Ookii.Dialogs.Wpf" version="1.1.0" targetFramework="net471" />
   <package id="SharpRemote" version="0.6.184" targetFramework="net471" />
